Create the org.cp.elements.data.struct package and introduce the Table API and data structure for modeling tabular data.
Create the Table interface modeling a tabular data structure.

Create the View interface modeling a particular image, or view of a Table, which is derived from a query defining the projection (selected columns), predicate (filtering the rows), order and so forth.

Create the Column and Row interfaces modeling the structure and contents of the Table respectively.