Modify the Javadoc comment on the View interface's getColumn(name:String) method to make note of the null return value if the Column does not exist in this View.
Add the getView() method to the Row interface to indicate the View to which the Row belongs.

Rename the getIndex() method to index().

Correct the spelling of the settValue(columnName:String, value:Object) method to setValue(..).

Add the getView() method to the Column interface to indicate the View to which the Column belongs.