Re-implement o.c.e.data.struct.SimpleBloomFilter in terms of the approximate/estimated number of elments stored by the Bloom Filter as well as the desired, acceptable false positive rate (probability).
Change the SimpleBloomFilter class type parameter to no longer be bound as a Number.

Add of(..) factory methods for constructing a SimpleBloomFilter based on approximate/estimated number of elements along with an optional desired, acceptable false positive rate.

Add formulas for calculating the required number of bits and optimal number of hash functions.

Refactor the SimpleBloomFilter constructor to initialize the filter with the number of bits and number of hash functions.

Add overridable method to calculate the bit array (filter) length based on the number of bits.

Refactor the add(..) and accept(..) method to seed the Random number generator with the hash code of the element.

Add a size() method to estimate the number of elements that have been added to the Bloom Filter.