Review, refactor and polish the o.c.e.data.struct.tabular.AbstractRow abstract class.
Add a factory method to construct a AbstractRow from an array of Object values.

Add getter and setter for the View to which this Row is associated.

Add a map(:Class<T>):T method to map the Row of values to an object of class type T.

Override to the toString() method to return the Row index.