Apply a precedence to the registered Converters in the ConversionService (ConverterRegistry) based on the Converter's declared order.
Fixes bug in the AbstractConversionService.convert(:Object, :Class<?>) method due to the random order of registered Converters in the Set of Converters registered in the ConversionService (ConverterRegistry) and which Converter gets applied.

This bug first appeared in Groovy / Gradle as of Gradle 7.5.1. The bug then appeared when upgrading to JDK 1.8.0_351, which made the Set ordering of the Converters in the ConversionService (ConverterRegistry) apparent.

Fixed buy by applying a consistent ordering using a List data structure combined with Elements' SmartComparator.